refactor(agents): move facet resolution, teardown, and fiber recovery into DynamicAgents
resolveSubAgent (the ctx.facets.get bootstrap + identity handshake),
abort/delete, recursive descendant destroy, workflow path invocation,
and the root-side facet fiber recovery scan now live on the
DynamicAgents class; the registry is owned by it. Agent keeps _cf_*
RPC entry points and public methods as delegates, plus a private
_runFacetInitInvocation helper so the module never touches the
invocation context machinery directly.
